+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+v0.1.6
+-------------
+Release Date: 2014-06-29
+
+  * Fixed issue when binding to Firebase limit() query
+  * Added error checking
+  * Added basic test suite
+  * Added distribution files to Firebase CDN
+
 v0.1.5
 -------------
 Release Date: 2014-05-23